alias 修正 lsのカラーオプション修正 プロンプト修正

+++ .bashrc
@@ -11,7 +11,7 @@ ulimit -c 0
 ## シェル変数
 # プロンプト文字列の設定
 #PS1="\n\d \t \w\e\n\u@\h:\$ "
-PS1="\n\w\e\n@\$"
+PS1='\[\033[0;34m\][\t \w]\n\[\033[0m\]\$ '
 # コマンド履歴のサイズ
 HISTSIZE=50000
 HISTFILESIZE=50000
@@ -36,20 +36,20 @@ export LS_COLORS
 #export LANG=ja_JP.Shift_JIS
 #export OUTPUT_CHARSET=sjis
 ## タイムゾーンの指定
-#export TZ=JST-9
+export TZ=JST-9
 ## デフォルトのページャをlessにする
-#export PAGER=less
+export PAGER=less
 ## lessのステータス行にファイル名と行数、％を表示する
-#export LESS='-X -r -i -P ?f%f:(stdin). ?lb%lb?L/%L.. [?eEOF:?pb%pb\%..]'
+export LESS='-X -r -i -P ?f%f:(stdin). ?lb%lb?L/%L.. [?eEOF:?pb%pb\%..]'
 ## lessの文字コードを指定する
 #export JLESSCHARSET=japanese-sjis
 ## DISPLAY変数を指定する
 #if [ -n ${DISPLAY} ]; then
-#	export DISPLAY=localhost:0.0
+#  export DISPLAY=localhost:0.0
 #fi
 ## TERM変数を指定する
 #if [ ! -n "${TERM}" ]; then
-#	export TERM=cygwin
+#  export TERM=cygwin
 #fi
 #
 #PATH=/usr/local/sbin:/usr/sbin:/sbin:.:~/bin
@@ -60,30 +60,31 @@ export LS_COLORS
 
 ## Alias 一覧
 # ls の8ビット透過にする
-alias ls='ls --color=auto'
+alias ls='ls -G'
 # updatedbでwindows領域を除外する
 alias updatedb='updatedb --localpaths="/bin /dev /etc /home /lib /sbin /srv /usr /var" --prunepaths="/usr/tmp /var/tmp"'
-alias la="ls -a"
+alias l="ls -a"
 alias ll="ls -l"
 alias j=jobs
 alias vi=vim
+alias c=cd
+alias u="cd ../"
+alias uu="cd ../../"
+alias uuu="cd ../../../"
 
-SYSTEM=`cygpath -u ${SYSTEMROOT}`/system32
-alias ping="$SYSTEM/ping.exe"
-alias arp="$SYSTEM/arp.exe"
-alias nslookup="$SYSTEM/nslookup.exe"
-alias traceroute="$SYSTEM/tracert.exe"
-alias route="$SYSTEM/route.exe"
-alias netstat="$SYSTEM/netstat.exe"
-alias ipconfig="$SYSTEM/ipconfig.exe"
-alias ifconfig=ipconfig
-alias start=cygstart
-
-alias fk="taskkill /f /im"
-alias apush="adb push"
-alias apull="adb pull"
-alias adev="adb devices"
-alias aver="adb version"
-alias akill="adb kill-server"
-
-unset SYSTEM                                  
+if [ -d cygpath ]; then
+    SYSTEM=`cygpath -u ${SYSTEMROOT}`/system32
+    alias ping="$SYSTEM/ping.exe"
+    alias arp="$SYSTEM/arp.exe"
+    alias nslookup="$SYSTEM/nslookup.exe"
+    alias traceroute="$SYSTEM/tracert.exe"
+    alias route="$SYSTEM/route.exe"
+    alias netstat="$SYSTEM/netstat.exe"
+    alias ipconfig="$SYSTEM/ipconfig.exe"
+    alias ifconfig=ipconfig
+    alias st="cygstart"
+    alias a="adb"
+    alias ad="adb devices"
+    alias fk="taskkill /f /im"
+    unset SYSTEM
+fi
